replacing old flatbuffer api by flap

+++ unity-space/Assets/src/game/ui/UIController.cs
@@ -15,7 +15,7 @@ namespace game.ui {
         public UIPrefabs prefabs;
 
         private bool hasSectorId;
-        private uint currentSectorId;
+        private ulong currentSectorId;
 
         public GameObject buttonsPanel;
         public GameObject sectorMapPanel;
